FastAPI Project/jwt/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
PyJWT==2.8.0
PyYAML==6.0.1
requests==2.31.0